+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/multiplatform/targets/HubdleKotlinMultiplatformNativeExtension.kt
@@ -52,7 +52,6 @@ constructor(
         configurable(priority = Priority.P6) {
             configure<KotlinMultiplatformExtension> {
                 val commonMain: KotlinSourceSet by sourceSets.getting
-                val iosArm32Main: KotlinSourceSet? = sourceSets.findByName("iosArm32Main")
                 val iosArm64Main: KotlinSourceSet? = sourceSets.findByName("iosArm64Main")
                 val iosX64Main: KotlinSourceSet? = sourceSets.findByName("iosX64Main")
                 val iosSimulatorArm64Main: KotlinSourceSet? =
@@ -78,7 +77,6 @@ constructor(
                     sourceSets.findByName("watchosSimulatorArm64Main")
 
                 val commonTest: KotlinSourceSet by sourceSets.getting
-                val iosArm32Test: KotlinSourceSet? = sourceSets.findByName("iosArm32Test")
                 val iosArm64Test: KotlinSourceSet? = sourceSets.findByName("iosArm64Test")
                 val iosX64Test: KotlinSourceSet? = sourceSets.findByName("iosX64Test")
                 val iosSimulatorArm64Test: KotlinSourceSet? =
@@ -105,7 +103,6 @@ constructor(
 
                 val nativeMainSourceSets: List<KotlinSourceSet> =
                     listOfNotNull(
-                        iosArm32Main,
                         iosArm64Main,
                         iosX64Main,
                         iosSimulatorArm64Main,
@@ -130,7 +127,6 @@ constructor(
 
                 val nativeTestSourceSets: List<KotlinSourceSet> =
                     listOfNotNull(
-                        iosArm32Test,
                         iosArm64Test,
                         iosX64Test,
                         iosSimulatorArm64Test,